Refactor AddMembershipModal
