+++ packages/ui/src/modals/AddMembershipModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Modal, ModalHeader } from '../../../components/Modal'
+import { Modal, ModalHeader } from '../components/Modal'
 
 interface MembershipModalProps {
   onClose: () => void
