+++ packages/ui/src/pages/Profile/MyMemberships/Memberships.tsx
@@ -1,7 +1,7 @@
 import { Membership } from '@joystream/types/members'
 import React from 'react'
 import styled from 'styled-components'
-import { CreateMembershipButton } from '../../../components/CreateMembershipButton'
+import { AddMembershipButton } from '../../../components/AddMembershipButton'
 import { Member } from './Member'
 
 const bob = ({
@@ -34,7 +34,7 @@ export function Memberships() {
           Amet minim mollit non deserunt ullamco est sit aliqua dolor do amet sint. Velit officia consequat duis enim
           velit mollit. Exercitation veniam consequat sunt nostrud amet.
         </p>
-        <CreateMembershipButton />
+        <AddMembershipButton />
       </>
     )
   }
